refactor: split mutable_id_tracker.rs into a module
Break the 1630-line file into a directory module with logically separated
components: change.rs (MappingChange + entry codec), mappings_storage.rs
(mappings file IO), versions_storage.rs (versions file IO), tests.rs (all
unit tests), and mod.rs (struct + IdTracker impl). No logic changes.
